core/src/test/java/org/gbif/pipelines/core/interpreter/taxonomy/TaxonomyInterpreterTest.java
package org.gbif.pipelines.core.interpreter.taxonomy;

import org.gbif.pipelines.core.utils.ExtendedRecordBuilder;
import org.gbif.pipelines.io.avro.ExtendedRecord;
import org.gbif.pipelines.io.avro.Rank;
import org.gbif.pipelines.io.avro.RankedName;

import java.util.HashMap;
import java.util.Map;
import java.util.stream.Collectors;

import org.junit.Assert;
import org.junit.Test;

/**
 * Tests the {@link TaxonomyInterpreter}.
 */
public class TaxonomyInterpreterTest {

  private static final String TEST_RECORD_ID = "testId";

  @Test
  public void testAssembledAuthor() {

    // Call to perform: https://api.gbif-uat.org/v1/species/match2?name=Puma
    // concolor&kingdom=Animalia&genus=Puma&rank=species
    ExtendedRecord record = new ExtendedRecordBuilder().kingdom("Animalia")
      .genus("Puma")
      .name("Puma concolor")
      .authorship("")
      .rank(Rank.SPECIES.name())
      .id(TEST_RECORD_ID)
      .build();

    InterpretedTaxonomy interpretedTaxon = null;
    try {

      try {
        interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);
      } catch (TaxonomyInterpretationException e) {
        e.printStackTrace();
      }

      Map<Rank, RankedName> ranksResponse = interpretedTaxon.getTaxonRecord()
        .getClassification()
        .stream()
        .collect(Collectors.toMap(rankedName -> rankedName.getRank(), rankedName -> rankedName));

      Assert.assertEquals(2435099, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(1, ranksResponse.get(Rank.KINGDOM).getKey().intValue());
      Assert.assertEquals("Chordata", ranksResponse.get(Rank.PHYLUM).getName());

      record = new ExtendedRecordBuilder().kingdom("Animalia")
        .genus("Puma")
        .name("Puma concolor (Linnaeus, 1771)")
        .rank(Rank.SPECIES.name())
        .id(TEST_RECORD_ID)
        .build();

      interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      ranksResponse = interpretedTaxon.getTaxonRecord()
        .getClassification()
        .stream()
        .collect(Collectors.toMap(rankedName -> rankedName.getRank(), rankedName -> rankedName));

      Assert.assertEquals(2435099, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(1, ranksResponse.get(Rank.KINGDOM).getKey().intValue());
      Assert.assertEquals("Chordata", ranksResponse.get(Rank.PHYLUM).getName());

      record = new ExtendedRecordBuilder().kingdom("Animalia")
        .genus("Puma")
        .name("Puma concolor")
        .authorship("(Linnaeus, 1771)")
        .rank(Rank.SPECIES.name())
        .id(TEST_RECORD_ID)
        .build();

      interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      ranksResponse = interpretedTaxon.getTaxonRecord()
        .getClassification()
        .stream()
        .collect(Collectors.toMap(rankedName -> rankedName.getRank(), rankedName -> rankedName));

      Assert.assertEquals(2435099, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(1, ranksResponse.get(Rank.KINGDOM).getKey().intValue());
      Assert.assertEquals("Chordata", ranksResponse.get(Rank.PHYLUM).getName());

    } catch (TaxonomyInterpretationException e) {
      Assert.fail(e.getMessage());
    }
  }

  @Test
  public void testOenanthe() {
    ExtendedRecord record = new ExtendedRecordBuilder().kingdom("Plantae")
      .name("Oenanthe")
      .authorship("")
      .rank(Rank.GENUS.name())
      .id(TEST_RECORD_ID)
      .build();

    try {

      InterpretedTaxonomy interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      Map<Rank, RankedName> ranksResponse = interpretedTaxon.getTaxonRecord()
        .getClassification()
        .stream()
        .collect(Collectors.toMap(rankedName -> rankedName.getRank(), rankedName -> rankedName));

      Assert.assertEquals(3034893, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(6, ranksResponse.get(Rank.KINGDOM).getKey().intValue());
      Assert.assertEquals("Oenanthe L.", interpretedTaxon.getTaxonRecord().getUsage().getName());

      record = new ExtendedRecordBuilder().kingdom("Plantae")
        .name("Oenanthe")
        .authorship("L.")
        .rank(Rank.GENUS.name())
        .id(TEST_RECORD_ID)
        .build();

      interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      ranksResponse = interpretedTaxon.getTaxonRecord()
        .getClassification()
        .stream()
        .collect(Collectors.toMap(rankedName -> rankedName.getRank(), rankedName -> rankedName));

      Assert.assertEquals(3034893, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(6, ranksResponse.get(Rank.KINGDOM).getKey().intValue());
      Assert.assertEquals("Oenanthe L.", interpretedTaxon.getTaxonRecord().getUsage().getName());

      record = new ExtendedRecordBuilder().kingdom("Animalia")
        .name("Oenanthe")
        .authorship("Vieillot, 1816")
        .rank(Rank.GENUS.name())
        .id(TEST_RECORD_ID)
        .build();

      interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      ranksResponse = interpretedTaxon.getTaxonRecord()
        .getClassification()
        .stream()
        .collect(Collectors.toMap(rankedName -> rankedName.getRank(), rankedName -> rankedName));

      Assert.assertEquals(2492483, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(1, ranksResponse.get(Rank.KINGDOM).getKey().intValue());
      Assert.assertEquals("Oenanthe Vieillot, 1816", interpretedTaxon.getTaxonRecord().getUsage().getName());

    } catch (TaxonomyInterpretationException e) {
      Assert.fail(e.getMessage());
    }
  }

  @Test
  public void testOtu() {
    ExtendedRecord record = new ExtendedRecordBuilder().kingdom("Animalia")
      .phylum("Annelida")
      .family("Lumbricidae")
      .name("BOLD:ACV7160")
      .rank(Rank.SPECIES.name())
      .id(TEST_RECORD_ID)
      .build();

    try {
      InterpretedTaxonomy interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      // FIXME: this tests does not pass. It is copied from old interpreter but it returns a different response.Should we check it??
      // this is the old test
//      Assert.assertEquals("BOLD:ACV7160", interpretedTaxon.getTaxonRecord().getUsage().getName());
      Assert.assertEquals("Lumbricidae", interpretedTaxon.getTaxonRecord().getUsage().getName());

    } catch (TaxonomyInterpretationException e) {
      Assert.fail(e.getMessage());
    }
  }

  @Test
  public void testCeratiaceae() {
    ExtendedRecord record = new ExtendedRecordBuilder().kingdom("Chromista")
      .phylum("Dinophyta")
      .clazz("Dinophyceae")
      .order("Peridiniales")
      .family("Ceratiaceae")
      .genus("Ceratium")
      .name("Ceratium hirundinella")
      .rank(Rank.SPECIES.name())
      .id(TEST_RECORD_ID)
      .build();

    try {
      InterpretedTaxonomy interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      Map<Rank, RankedName> ranksResponse = interpretedTaxon.getTaxonRecord()
        .getClassification()
        .stream()
        .collect(Collectors.toMap(rankedName -> rankedName.getRank(), rankedName -> rankedName));

      Assert.assertEquals(7598904, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(7479242, ranksResponse.get(Rank.FAMILY).getKey().intValue());
      // FIXME: this tests does not pass. It is copied from old interpreter but it returns a different response.Should we check it??
      // this is the old test
//      Assert.assertEquals("Ceratium hirundinella (O.F.Müller) Dujardin, 1841",
//                          interpretedTaxon.getTaxonRecord().getUsage().getName());
      Assert.assertEquals("Ceratium hirundinella (O.F.Müll.) Dujard.",
                          interpretedTaxon.getTaxonRecord().getUsage().getName());

    } catch (TaxonomyInterpretationException e) {
      Assert.fail(e.getMessage());
    }

  }

  @Test
  public void testNubLookupGood() {
    ExtendedRecord record = new ExtendedRecordBuilder().kingdom("Animalia")
      .genus("Puma")
      .name("Puma concolor")
      .rank(Rank.SPECIES.name())
      .id(TEST_RECORD_ID)
      .build();

    try {
      InterpretedTaxonomy interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      Map<Rank, RankedName> ranksResponse = interpretedTaxon.getTaxonRecord()
        .getClassification()
        .stream()
        .collect(Collectors.toMap(rankedName -> rankedName.getRank(), rankedName -> rankedName));

      Assert.assertEquals(2435099, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(1, ranksResponse.get(Rank.KINGDOM).getKey().intValue());
      Assert.assertEquals("Chordata", ranksResponse.get(Rank.PHYLUM).getName());

    } catch (TaxonomyInterpretationException e) {
      Assert.fail(e.getMessage());
    }

  }

  @Test
  public void testAcceptedUsage() {
    ExtendedRecord record = new ExtendedRecordBuilder().name("Agallisus lepturoides").id(TEST_RECORD_ID).build();

    try {
      InterpretedTaxonomy interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);

      Assert.assertEquals(1118030, interpretedTaxon.getTaxonRecord().getUsage().getKey().intValue());
      Assert.assertEquals(1118026, interpretedTaxon.getTaxonRecord().getAcceptedUsage().getKey().intValue());
      Assert.assertEquals("Agallisus lepturoides Hovore, Penrose & Neck, 1987",
                          interpretedTaxon.getTaxonRecord().getUsage().getName());
      Assert.assertEquals("Agallissus lepturoides (Chevrolat, 1849)",
                          interpretedTaxon.getTaxonRecord().getAcceptedUsage().getName());

    } catch (TaxonomyInterpretationException e) {
      Assert.fail(e.getMessage());
    }

  }

  @Test(expected = TaxonomyInterpretationException.class)
  public void testMatchNullArgs() throws TaxonomyInterpretationException {
    InterpretedTaxonomy interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(new ExtendedRecord());
  }

  @Test(expected = TaxonomyInterpretationException.class)
  public void testMatchEmptyArgs() throws TaxonomyInterpretationException {
    ExtendedRecord record = new ExtendedRecord();
    record.setCoreTerms(new HashMap<>());
    InterpretedTaxonomy interpretedTaxon = TaxonomyInterpreter.interpretTaxonomyFields(record);
  }

}
